Revert "Этап 2. Основной сервис"
This reverts commit 87bf061d

+++ stats/stats-dto/src/main/java/ru/practicum/hit/dto/HitDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class HitDto {
     String app;
     String uri;
-    Long hits;
+    Integer hits;
 }
